Refactor EcMath to use ModularArithmetic, also EcDH now works with Messages
